refactor: 7/18 코드리뷰 반영 (상수 클래스 분리)
