+++ src/main/java/gift/util/constants/CategoryConstants.java
@@ -1,0 +1,8 @@
+package gift.util.constants;
+
+public class CategoryConstants {
+
+    public static final String CATEGORY_NOT_FOUND = "카테고리를 다음의 id로 찾을 수 없습니다. id: ";
+    public static final String INVALID_COLOR = "색상 코드 형식이 옳지 않습니다.";
+    public static final String CATEGORY_DESCRIPTION_SIZE_LIMIT = "카테고리 설명은 공백을 포함하여 최대 255자까지 입력할 수 있습니다.";
+}
