+++ src/main/java/gift/util/constants/MemberConstants.java
@@ -1,0 +1,11 @@
+package gift.util.constants;
+
+public class MemberConstants {
+
+    public static final String INVALID_CREDENTIALS = "이메일 또는 비밀번호가 잘못되었습니다.";
+    public static final String EMAIL_ALREADY_USED = "이미 사용 중인 이메일입니다.";
+    public static final String ID_NOT_FOUND = "존재하지 않는 ID입니다.";
+    public static final String INVALID_AUTHORIZATION_HEADER = "유효하지 않은 Authorization 헤더입니다.";
+    public static final String INVALID_EMAIL = "유효한 이메일 주소를 입력해야 합니다.";
+    public static final String INVALID_PASSWORD = "비밀번호는 최소 4자리 이상이어야 합니다.";
+}
