+++ src/main/java/gift/util/constants/ProductConstants.java
@@ -1,0 +1,11 @@
+package gift.util.constants;
+
+public class ProductConstants {
+
+    public static final String PRODUCT_NOT_FOUND = "상품을 다음의 id로 찾을 수 없습니다. id: ";
+    public static final String INVALID_PRICE = "가격은 0 이상으로 설정되어야 합니다.";
+    public static final String REQUIRED_FIELD_MISSING = "필수 입력 사항이 누락되었습니다.";
+    public static final String PRODUCT_NAME_SIZE_LIMIT = "상품 이름은 공백을 포함하여 최대 15자까지 입력할 수 있습니다.";
+    public static final String PRODUCT_NAME_INVALID_CHARACTERS = "상품 이름에는 다음 특수 문자의 사용만 허용됩니다: ( ), [ ], +, -, &, /, _";
+    public static final String PRODUCT_NAME_REQUIRES_APPROVAL = "\"카카오\"가 포함된 문구는 담당 MD와 협의한 경우에만 사용할 수 있습니다.";
+}
