+++ src/main/java/gift/util/constants/WishConstants.java
@@ -1,0 +1,8 @@
+package gift.util.constants;
+
+public class WishConstants {
+
+    public static final String WISH_NOT_FOUND = "위시를 다음의 id로 찾을 수 없습니다. id: ";
+    public static final String WISH_ALREADY_EXISTS = "이미 위시 리스트에 추가된 상품입니다.";
+    public static final String PERMISSION_DENIED = "삭제 권한이 없습니다.";
+}
