+++ src/test/java/gift/model/ProductTest.java
@@ -1,8 +1,8 @@
 package gift.model;
 
-import static gift.util.Constants.PRODUCT_NAME_INVALID_CHARACTERS;
-import static gift.util.Constants.PRODUCT_NAME_REQUIRES_APPROVAL;
-import static gift.util.Constants.PRODUCT_NAME_SIZE_LIMIT;
+import static gift.util.constants.ProductConstants.PRODUCT_NAME_INVALID_CHARACTERS;
+import static gift.util.constants.ProductConstants.PRODUCT_NAME_REQUIRES_APPROVAL;
+import static gift.util.constants.ProductConstants.PRODUCT_NAME_SIZE_LIMIT;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import jakarta.validation.ConstraintViolation;
