+++ src/test/java/gift/service/CategoryServiceTest.java
@@ -1,6 +1,6 @@
 package gift.service;
 
-import static gift.util.Constants.CATEGORY_NOT_FOUND;
+import static gift.util.constants.CategoryConstants.CATEGORY_NOT_FOUND;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.when;
